# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pets App API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/pins/{pin}/reviews": {
            "get": {
                "tags": [
                    "Pins",
                    "Reviews"
                ],
                "summary": "Get list of the reviews of the pin.",
                "operationId": "48e86af93bffc91fe3fcc340470d7ab8",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewFullResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pins",
                    "Reviews"
                ],
                "summary": "Add review to the pin.",
                "operationId": "6d40c04c2d7566359d8fb2bf6afc3b7d",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewFullResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pins/{pin}/reviews/{review}": {
            "delete": {
                "tags": [
                    "Pins",
                    "Reviews"
                ],
                "summary": "Delete a pin.",
                "operationId": "b0fdeff530bf34ee65bd58ceeba9ce73",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824cb"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Pins",
                    "Reviews"
                ],
                "summary": "Update pin review.",
                "operationId": "6ab53c18f72fb2d569dedb2707a23b21",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824cb"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewFullResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pins/{pin}/reviews/{review}/media": {
            "get": {
                "tags": [
                    "Pins"
                ],
                "summary": "Get list of the review media.",
                "operationId": "db9f3361a4ee275a2d8f48d84957f33a",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824cb"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediaShortResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pins"
                ],
                "summary": "Upload review media.",
                "operationId": "8d279a513d9aa68ee9d0db3a7e0073da",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824cb"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PinMediaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MediaShortResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pins/{pin}/reviews/{review}/media/{media}": {
            "delete": {
                "tags": [
                    "Pins"
                ],
                "summary": "Delete a pin.",
                "operationId": "0796a867b71e417899e32f292a357ea3",
                "parameters": [
                    {
                        "name": "pin",
                        "in": "path",
                        "description": "UUID of pin.",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "review",
                        "in": "path",
                        "description": "UUID of pin.",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824ca"
                    },
                    {
                        "name": "media",
                        "in": "path",
                        "description": "UUID of media which should be removed.",
                        "required": true,
                        "example": "995037a6-60b3-4055-aa14-3513aa9824cc"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PinMediaRequest": {
                "properties": {
                    "files": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        }
                    }
                },
                "type": "object"
            },
            "ReviewStoreRequest": {
                "properties": {
                    "rating": {
                        "type": "integer",
                        "example": 5
                    },
                    "message": {
                        "type": "string",
                        "example": "Some message.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MediaShortResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PinReviewResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    },
                    "reviewable": {
                        "$ref": "#/components/schemas/UserShortResource"
                    }
                },
                "type": "object"
            },
...
            "ReviewFullResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "pin": {
                        "$ref": "#/components/schemas/PinReviewResource"
                    }
                },
                "type": "object"
            },
...
            "UserShortResource": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}